2-HFMCP-python/requirements.txt
# === Core libraries ===
mcp>=1.2.0           # Official Model Context Protocol SDK for Python
openai>=1.42.0       # Azure OpenAI / OpenAI SDK with tool-calling support

# === Optional helpers (recommended for local testing) ===
httpx>=0.27.0        # HTTP client used internally by MCP (async)
aiohttp>=3.10.0      # For alternative SSE transport (if needed)
python-dotenv>=1.0.1 # For managing .env environment variables easily
